Fixed compilation warnings and updated to version 1.0.0-SNAPSHOT

+++ src/main/scala/scodec/stream/codec/StreamCodec.scala
@@ -3,7 +3,7 @@ package codec
 
 trait StreamCodec[A] extends StreamGenCodec[A,A] {
 
-  override def take(n: Int): StreamCodec[A] =
+  override def take(n: Long): StreamCodec[A] =
     StreamCodec.instance(StreamEncoder.instance(encoder).take(n),
                          StreamDecoder.instance(decoder).take(n))
